requirements.txt
# ⚠️ Python 3.8 相容版本
# Web Scraping
playwright>=1.40.0,<1.49.0

# LLM (openai>=1.56 需要 Python 3.9+，請固定在此版本以下)
openai>=1.12.0,<1.56.0

# Data Storage
aiosqlite>=0.19.0

# Report Generation
Jinja2>=3.1.3

# Notification (telegram 20.x 支援 Python 3.8，httpx 需固定 ~0.26)
python-telegram-bot>=20.0,<21.0
httpx~=0.26.0

# Scheduling
APScheduler>=3.10.0

# Configuration
python-dotenv>=1.0.0

# Utilities
rich>=13.0.0
